Update to Kubernetes v1.19.0
kube-proxy container has no longer read-only file system, as when it
syncs iptable rules, it runs the script, which runs update-alternatives,
which writes to /etc/alternatives and /var/log to update which iptables
binaries should be used.

I've tried putting emptyDir volumes for those locations, but then it
creates symlinks between the mounts, so the binaries are not found and
cannot be executed.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>